requirements.txt
APScheduler==3.11.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.0
geographiclib==2.0
gevent
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
loguru==0.7.3
MarkupSafe==3.0.2
msgspec==0.19.0
names-dataset==3.1.0
nltk==3.9.1
pycountry==24.6.1
pycparser==2.22
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
simple-websocket==1.1.0
tqdm==4.67.1
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
